Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
PyMuPDF
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
langdetect==1.0.9
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.3.3
jsonschema==4.23.0
